little change

+++ NMP-Portal/NMP-Commons/Models/SoilAnalysis.cs
@@ -48,7 +48,7 @@ public class SoilAnalysis
     public string? MagnesiumAnalysis { get; set; }
     public string? MagnesiumStatus { get; set; }
     public string? NitrogenResidueGroup { get; set; }
-    public decimal? OrganicMatterPercentage { get; set; }
+    public int? OrganicMatterPercentage { get; set; }
     public string? Comments { get; set; }
     public int? PreviousID { get; set; }
     public DateTime? CreatedOn { get; set; }
